third_party/blink/web_tests/platform/generic/fast/xpath/4XPath/Core/test_parser-expected.txt

PASS child::*
PASS /child::*
PASS /*/*
PASS /child::*/*/child::GCHILD
PASS //*
PASS //GCHILD
PASS //@attr1
PASS x:GCHILD
PASS .//GCHILD
PASS .//GCHILD
PASS /
PASS //CHILD1/..
PASS .//foo:*
PASS CHILD1 | CHILD2
PASS descendant::GCHILD[3]
PASS descendant::GCHILD[parent::CHILD1]
PASS descendant::GCHILD[position() > 1]
PASS @attr1[.='val1']
PASS DOM.evaluate("1", ROOT, null, XPathResult.ANY_TYPE, null).numberValue is 1
PASS DOM.evaluate("00200", ROOT, null, XPathResult.ANY_TYPE, null).numberValue is 200
PASS DOM.evaluate("3+4*7", ROOT, null, XPathResult.ANY_TYPE, null).numberValue is 31
PASS DOM.evaluate("3-4*1", ROOT, null, XPathResult.ANY_TYPE, null).numberValue is -1
PASS DOM.evaluate("string('1')", ROOT, null, XPathResult.ANY_TYPE, null).stringValue is '1'
PASS DOM.evaluate("concat('1', '2')", ROOT, null, XPathResult.ANY_TYPE, null).stringValue is '12'
PASS DOM.evaluate("true()", ROOT, null, XPathResult.ANY_TYPE, null).booleanValue is true
PASS DOM.evaluate("false()", ROOT, null, XPathResult.ANY_TYPE, null).booleanValue is false
PASS DOM.evaluate("1=3<4", ROOT, null, XPathResult.ANY_TYPE, null).booleanValue is true
PASS DOM.evaluate("1 or 2 and 3", ROOT, null, XPathResult.ANY_TYPE, null).booleanValue is true
PASS DOM.evaluate("1 and 2 = 3", ROOT, null, XPathResult.ANY_TYPE, null).booleanValue is false
PASS DOM.evaluate("-1 or 2", ROOT, null, XPathResult.ANY_TYPE, null).booleanValue is true
PASS DOM.evaluate(". or *", ROOT, null, XPathResult.ANY_TYPE, null).booleanValue is true
PASS text()
PASS processing-instruction('f')
PASS (child::*)
PASS DOM.evaluate(". * 0", ROOT, null, XPathResult.NUMBER_TYPE, null).numberValue is NaN
PASS DOM.evaluate(".. * 0", ROOT, null, XPathResult.NUMBER_TYPE, null).numberValue is NaN
PASS DOM.evaluate("/.. * 0", ROOT, null, XPathResult.NUMBER_TYPE, null).numberValue is NaN
PASS CHILD2/@CODE
PASS DOM.evaluate("CHILD2/@CODE * 0", ROOT, null, XPathResult.NUMBER_TYPE, null).numberValue is 0
PASS föø
PASS DOM.evaluate("\\", ROOT, null, XPathResult.ORDERED_NODE_SNAPSHOT_TYPE, null) threw exception SyntaxError: Failed to execute 'evaluate' on 'Document': The string '\' is not a valid XPath expression..
PASS successfullyParsed is true

TEST COMPLETE
